exec: widen the inequality-join range axis to 128-bit for decimal and mixed-numeric keys (#876)
* exec: widen the inequality-join range axis to i128 for decimal and mixed-numeric keys

A combine whose range/inequality predicate compared decimal columns — or mixed
an integer and a float across the two inputs — silently produced empty or wrong
output when the planner selected the block-band inequality join. The numeric
range axis reduced each key to an i64: decimals fell through to "unrepresentable"
and routed every row out of the match, and an integer versus a float were encoded
in incompatible spaces on the same axis, so the bounds prune and the kernel
compared them wrongly.

Widen the range axis from i64 to i128 end to end — the block bounds and prune,
the sort payloads, the drain/sort/spill keys, and the IEJoin/PWMJ kernels — and
pick each axis's order-preserving reduction from the typechecker's unified
operand type:

- integers widen directly; finite floats take the existing orderable-float
  transform; dates and datetimes keep their epoch offset. These stay
  byte-identical in the join result to before, so existing int/float/date/
  datetime range joins are unchanged.
- a mixed integer/float axis reduces both sides through the float transform so
  the two encodings agree — the integer is compared as a float, matching how the
  comparison operators treat it elsewhere.
- decimals are placed on a shared 10^18 fixed-point grid (mantissa scaled by
  10^(18 - scale)) as an i128 — exact and scale-invariant, so 2.5 and 2.50
  compare equal and a monetary band join matches correctly.

A decimal range value that cannot be represented exactly — a magnitude that
overflows the grid, or more than 18 fractional digits, which would truncate — is
no longer dropped or lossily projected: the run fails loud with a new typed
diagnostic E326 naming the combine, so a wrong or empty result is never emitted.
Integer, float, and mixed integer/float range keys are always representable.

Tests: a decimal band-join oracle and a mixed integer/float oracle, each checked
against a brute-force reference over the real typed values across a spilling and
a resident memory budget (byte-identical across budgets); the overflow case
asserting the fail-loud E326; the kernel property tests widened to i128; the
existing range-join fixtures unchanged. Docs: the combine user guide, the engine
internals, and the E326 explain page.

* exec: close the SortMerge decimal/mixed gap and fail loud on unusable range keys

Follow-up hardening for the i128 range-axis widening.

SortMerge was a surviving pre-widening path: its kernel maps each range Value to a
monotone i64 WITHIN its own type (decimals to a constant, int and float in disjoint
per-type spaces), and its eligibility gate never consulted the operand types. A
single-inequality decimal or mixed int/float range join on pre-sorted inputs would
route there and silently drop or mismatch. Gate SortMerge eligibility on the range
axis's key type: keep it only for a same-type Integer/Float/Date/DateTime pair,
where its per-type encoding is consistent, and route a decimal or mixed int/float
axis to the IEJoin/block-band path, which reduces both sides to the one i128 space.
The result is identical either way; only the physical strategy differs.

Classify the range axis from the CONCRETE operand types rather than the
typechecker's unified type, so a mixed int/float pair is distinguished from a pure
float pair (they need different handling on the SortMerge path). Two operand shapes
cannot back an exact axis and are now rejected at plan time with a new E327 rather
than routed to the axis where they would drop rows or mismatch: an ambiguous
`numeric` operand — the result of abs/min/max/clamp, whose runtime value may be an
integer compared exactly OR a float, so no single reduction agrees with the exact
comparison for every value — and a non-orderable pairing such as a string
comparison. This also replaces the earlier silent route-out of a non-orderable
range key with a fail-loud plan diagnostic.

Also:
- E326 (a decimal range value beyond the fixed-point axis) is a bad-data-value
  halt, the same class as an eval failure — move it from exit 1 to exit 3.
- Normalize a decimal's trailing fractional zeros before the 18-digit truncation
  check, so an exactly-representable value padded past scale 18 is not a false E326.
- Widen the kernel property-test generators to draw values that straddle the i64
  limits and span the full i128 range, so the widened comparators are fuzzed there.

Tests: a presorted decimal and a presorted mixed int/float single-inequality range
join reach the SortMerge eligibility path and route to IEJoin with correct oracle
results, while a presorted same-type int join still uses SortMerge; abs/numeric,
decimal-vs-numeric, and string range operands are each rejected with E327; the
trailing-zero decimal normalizes; and the kernel proptests pass over the widened
range. Docs: the combine user guide, the engine internals, and a new E327 page.

+++ crates/clinker-exec/src/pipeline/iejoin/block.rs
@@ -296,8 +296,8 @@ fn pair_bytes<P: HeapBytes>(record: &Record, payload: &P) -> usize {
 /// columns at this width, so `Single`'s dead second axis is never counted.
 fn range_key_width(op2: Option<RangeOp>) -> usize {
     match op2 {
-        Some(_) => std::mem::size_of::<(i64, i64)>(),
-        None => std::mem::size_of::<i64>(),
+        Some(_) => std::mem::size_of::<(i128, i128)>(),
+        None => std::mem::size_of::<i128>(),
     }
 }
 
@@ -358,20 +358,20 @@ fn eq_match(driver_eq: &[u8], build_eq: &[u8]) -> bool {
 enum RangeMode {
     Dual {
         op2: RangeOp,
-        driver_keys: Vec<(i64, i64)>,
+        driver_keys: Vec<(i128, i128)>,
     },
     Single {
-        driver_keys: Vec<i64>,
+        driver_keys: Vec<i128>,
     },
 }
 
 /// Per-axis key bounds accumulated while a block fills.
 #[derive(Clone, Copy)]
 struct Bounds {
-    k1_min: i64,
-    k1_max: i64,
-    k2_min: i64,
-    k2_max: i64,
+    k1_min: i128,
+    k1_max: i128,
+    k2_min: i128,
+    k2_max: i128,
 }
 
 impl Bounds {
@@ -379,14 +379,14 @@ impl Bounds {
     /// an unfilled block (never updated) prunes against every operator.
     fn empty() -> Self {
         Self {
-            k1_min: i64::MAX,
-            k1_max: i64::MIN,
-            k2_min: i64::MAX,
-            k2_max: i64::MIN,
+            k1_min: i128::MAX,
+            k1_max: i128::MIN,
+            k2_min: i128::MAX,
+            k2_max: i128::MIN,
         }
     }
 
-    fn update(&mut self, k1: i64, k2: i64) {
+    fn update(&mut self, k1: i128, k2: i128) {
         self.k1_min = self.k1_min.min(k1);
         self.k1_max = self.k1_max.max(k1);
         self.k2_min = self.k2_min.min(k2);
@@ -483,7 +483,7 @@ impl<P: DeserializeOwned> Block<P> {
 /// `op` is already driver-vs-build oriented (the executor flips the raw
 /// predicate before the kernel), matching `iejoin_numeric`'s `left OP right`
 /// convention.
-fn possible(op: RangeOp, dmin: i64, dmax: i64, bmin: i64, bmax: i64) -> bool {
+fn possible(op: RangeOp, dmin: i128, dmax: i128, bmin: i128, bmax: i128) -> bool {
     match op {
         // ∃ d,b: d < b  ⇔  min d < max b
         RangeOp::Lt => dmin < bmax,
@@ -500,7 +500,7 @@ fn possible(op: RangeOp, dmin: i64, dmax: i64, bmin: i64, bmax: i64) -> bool {
 /// `pwmj_numeric` would return for the same keys — so the two paths emit the
 /// identical match set for a block-pair.
 #[inline]
-fn apply_range(op: RangeOp, a: i64, b: i64) -> bool {
+fn apply_range(op: RangeOp, a: i128, b: i128) -> bool {
     match op {
         RangeOp::Lt => a < b,
         RangeOp::Le => a <= b,
@@ -1076,12 +1076,12 @@ pub(super) fn execute_block_band(
                 (budget.hard_limit().saturating_sub(pair_peak_reserved) / (3 * pair_size)) as usize;
             let materialized = match &range_mode {
                 RangeMode::Dual { op2, driver_keys } => {
-                    let bkeys: Vec<(i64, i64)> =
+                    let bkeys: Vec<(i128, i128)> =
                         build_loaded.iter().map(|(_, p)| (p.k1, p.k2)).collect();
                     iejoin_numeric_capped(driver_keys, &bkeys, op1, *op2, max_pairs)
                 }
                 RangeMode::Single { driver_keys } => {
-                    let bl: Vec<i64> = build_loaded.iter().map(|(_, p)| p.k1).collect();
+                    let bl: Vec<i128> = build_loaded.iter().map(|(_, p)| p.k1).collect();
                     pwmj_numeric_capped(driver_keys, &bl, op1, max_pairs)
                 }
             };
@@ -1438,8 +1438,8 @@ fn spilled_scratch_bytes<P>(block: &Block<P>) -> u64 {
 #[derive(Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
 struct BuildPayload {
     eq_hash: u64,
-    k1: i64,
-    k2: i64,
+    k1: i128,
+    k2: i128,
     build_idx: u64,
     eq: Vec<u8>,
 }
@@ -1454,8 +1454,8 @@ struct BuildPayload {
 #[derive(Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
 struct DriverPayload {
     eq_hash: u64,
-    k1: i64,
-    k2: i64,
+    k1: i128,
+    k2: i128,
     driver_idx: u64,
     order: RecordOrder,
     eq: Vec<u8>,
@@ -1664,7 +1664,7 @@ fn finish_and_slice<P: Serialize + DeserializeOwned + Send + Ord + HeapBytes>(
     schema: &Arc<Schema>,
     ctx: &DrainCtx<'_>,
     resident: &mut ResidentBudget,
-    key_of: &impl Fn(&P) -> (u64, i64, i64),
+    key_of: &impl Fn(&P) -> (u64, i128, i128),
 ) -> Result<Vec<Block<P>>, PipelineError> {
     let pre_finish = buf.bytes_used() as u64;
     let (sorted, residue) = buf.finish().map_err(|e| {
@@ -1725,7 +1725,7 @@ fn slice_side<P: Serialize + DeserializeOwned + Send + Ord + HeapBytes>(
     schema: &Arc<Schema>,
     ctx: &DrainCtx<'_>,
     resident: &mut ResidentBudget,
-    key_of: &impl Fn(&P) -> (u64, i64, i64),
+    key_of: &impl Fn(&P) -> (u64, i128, i128),
 ) -> Result<Vec<Block<P>>, PipelineError> {
     let mut blocks: Vec<Block<P>> = Vec::new();
     let mut cur: Vec<(Record, P)> = Vec::new();
@@ -2230,7 +2230,7 @@ mod tests {
             .map(|(key, _)| match key {
                 Some((k1, k2)) => RecordScan::Matched {
                     eq_hash: 0,
-                    range_key: (*k1, *k2),
+                    range_key: (*k1 as i128, *k2 as i128),
                     eq: Vec::new(),
                 },
                 None => RecordScan::Unmatched,
@@ -2427,12 +2427,12 @@ mod tests {
         // Driver bounds fixed at [10, 20]; build bounds chosen so each axis
         // configuration is disjoint-below, touching-equal, overlapping, or
         // disjoint-above relative to the driver.
-        let (dmin, dmax) = (10i64, 20i64);
+        let (dmin, dmax) = (10i128, 20i128);
         // (bmin, bmax, label) fixtures.
-        let disjoint_below = (0i64, 5i64); // build strictly below driver
-        let touching_equal = (20i64, 30i64); // build min == driver max
-        let overlapping = (15i64, 25i64); // ranges overlap
-        let disjoint_above = (25i64, 40i64); // build strictly above driver
+        let disjoint_below = (0i128, 5i128); // build strictly below driver
+        let touching_equal = (20i128, 30i128); // build min == driver max
+        let overlapping = (15i128, 25i128); // ranges overlap
+        let disjoint_above = (25i128, 40i128); // build strictly above driver
 
         // Lt: possible ⇔ dmin < bmax.
         assert!(!possible(
@@ -3830,7 +3830,7 @@ mod tests {
                 .iter()
                 .map(|(k1, k2, _, _)| RecordScan::Matched {
                     eq_hash: 0,
-                    range_key: (*k1, *k2),
+                    range_key: (*k1 as i128, *k2 as i128),
                     eq: Vec::new(),
                 })
                 .collect();
@@ -3924,7 +3924,7 @@ mod tests {
                 .iter()
                 .map(|(k1, k2, _, _)| RecordScan::Matched {
                     eq_hash: 0,
-                    range_key: (*k1, *k2),
+                    range_key: (*k1 as i128, *k2 as i128),
                     eq: Vec::new(),
                 })
                 .collect();
@@ -4026,7 +4026,7 @@ mod tests {
         let driver_scans: Vec<RecordScan> = (0..n)
             .map(|i| RecordScan::Matched {
                 eq_hash: 0,
-                range_key: (i, i),
+                range_key: (i as i128, i as i128),
                 eq: Vec::new(),
             })
             .collect();
@@ -4046,7 +4046,7 @@ mod tests {
         let build_scans: Vec<RecordScan> = (0..n)
             .map(|i| RecordScan::Matched {
                 eq_hash: 0,
-                range_key: (i, i),
+                range_key: (i as i128, i as i128),
                 eq: Vec::new(),
             })
             .collect();
@@ -4958,7 +4958,7 @@ mod tests {
                 Some((k1, k2)) => RecordScan::Matched {
                     eq_hash: hash_of(r.group),
                     eq: eq_of(r.group),
-                    range_key: (k1, k2),
+                    range_key: (k1 as i128, k2 as i128),
                 },
                 None => RecordScan::Unmatched,
             })
